save commit with vision

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/data/transfer.java
@@ -2,6 +2,6 @@ package org.firstinspires.ftc.teamcode.data;
 
 public class transfer {
     public static double armRotationPos = 0;
-    public static double armExtensionPos = 0;
+    public static int armExtensionPos = 0;
     public static double angle = 0;
 }
